+++ admin/includes/sidebar.php
@@ -7,7 +7,8 @@
         <li><a href="dashboard.php">Dashboard</a></li>
 
         <li class="sidebar-separator"></li>
-        <li><a href="pages.php">Pages</a></li>
+        <li><a href="content.php?type=post">Posts</a></li>
+        <li><a href="content.php?type=page">Pages</a></li>
         <li><a href="menu.php">Menu</a></li>
         <li><a href="media.php">Media</a></li>