+++ admin/new.php
@@ -11,29 +11,40 @@ $temppass = $_SESSION['temppass'];
 $houndAdmin = new hound('', '');
 $param = $houndAdmin->read_param('../site/config.txt');
 
-if($temppass == $password) {
+if ((string) $temppass === (string) $password) {
     include 'includes/header.php';
     include 'includes/sidebar.php'; ?>
 
     <div class="content">
         <div class="content main">
             <?php
-            if($_POST['op'] == 'mod') {
-                $slug=$_POST['slug'];
-                $title=$_POST['title'];
-                $content=$_POST['content'];
-                $content=str_replace("\n","",$content);
-                $content=str_replace("\\","",$content);
+			$type = houndSanitizeString($_GET['type']);
+			$acceptedTypes = array('post', 'page');
 
-                $metatitle=$_POST['metatitle'];
-                $metadescription=$_POST['metadescription'];
+			if (!in_array($type, $acceptedTypes)) {
+				$type = 'page';
 
-                if(strlen($metatitle)<=0)$metatitle="$title";
-                if(strlen($metadescription)<=0)$metadescription= strip_tags(substr($content,0,160));
+				echo '<div class="thin-ui-notification thin-ui-notification-error">Invalid item type. Switching to page type.</div>';
+			}
 
-                $template=$_POST['template'];
+            if ((string) $_POST['op'] === 'mod') {
+                $slug = $_POST['slug'];
+                $title = $_POST['title'];
+                $content = $_POST['content'];
+                $content = str_replace("\n", "", $content);
+                $content = str_replace("\\", "", $content);
 
-                $file = '../site/pages/page-' . $slug . '.txt';
+                $metatitle = $_POST['metatitle'];
+                $metadescription = $_POST['metadescription'];
+
+                if (strlen($metatitle) <= 0)
+                    $metatitle = "$title";
+                if (strlen($metadescription) <= 0)
+                    $metadescription = strip_tags(substr($content, 0, 160));
+
+                $template = $_POST['template'];
+
+                $file = '../site/pages/' . $type . '-' . $slug . '.txt';
                 echo '<div class="thin-ui-notification thin-ui-notification-success">Created <code>' . $file . '</code>.</div>';
 
                 //create file
@@ -48,56 +59,40 @@ if($temppass == $password) {
                     'Meta.description' => $metadescription,
                     'Template'=> $template
                 );
-                //print_r($arrayvalue);
-                if(writeParam($arrayvalue,$file))echo "
-                <div class=\"panel panel-success\">
-                    <div class=\"panel-heading\">
-                      <h3 class=\"panel-title\">Success</h3>
-                    </div>
-                    <div class=\"panel-body\">
-                      Updated
-                    </div>
-                  </div>
-                ";
-                else echo "
-                  <div class=\"panel panel-error\">
-                    <div class=\"panel-heading\">
-                      <h3 class=\"panel-title\">Error</h3>
-                    </div>
-                    <div class=\"panel-body\">
-                      I/o error
-                    </div>
-                  </div>
-                ";
+                if (writeParam($arrayvalue, $file)) {
+                    echo '<div class="thin-ui-notification thin-ui-notification-success">Changes saved successfully.</div>';
+                } else {
+                    echo '<div class="thin-ui-notification thin-ui-notification-error">An error occurred while saving changes.</div>';
+                }
             }
             ?>
 
-            <h2>New page</h2>
+            <h2>New <?php echo $type; ?></h2>
 
-            <form role="form" name="form1" id="form1" action="new-page.php" method="post">
+            <form role="form" name="form1" id="form1" action="new.php" method="post">
                 <input type="hidden" value="mod" name="op">
 
                 <p>
                     <b>Title</b><br>
                     <input onkeyup="houndSlugify('title', 'slug');" id="title" name="title" required type="text" class="thin-ui-input" size="64">
-                    <br><small>Title of page</small>
+                    <br><small>Title of <?php echo $type; ?></small>
                 </p>
 
                 <p>
                     <b>Slug</b><br>
                     <input name="slug" id="slug" required type="text" class="thin-ui-input" size="64">
-                    <br><small>A unique page identification string</small>
+                    <br><small>A unique <?php echo $type; ?> identification string</small>
                 </p>
 
                 <p>
                     <b>Content</b><br>
                     <textarea id="txtTextArea1" name="content" style="width:100%" rows="20"></textarea>
-                    <br><small>Content of page</small>
+                    <br><small>Content of <?php echo $type; ?></small>
                 </p>
 
                 <p>
                     <b>Template</b>
-                    <br><small>Template of page</small>
+                    <br><small>Template of <?php echo $type; ?></small>
                     <div class="thin-ui-select-wrapper">
                         <select name="template" id="template">
                             <?php
@@ -127,7 +122,7 @@ if($temppass == $password) {
                 </p>
 
                 <p>
-                    <button type="submit" class="thin-ui-button thin-ui-button-primary">Create page</button>
+                    <button type="submit" class="thin-ui-button thin-ui-button-primary">Create <?php echo $type; ?></button>
                 </p>
             </form>
         </div>